Porting to Redmine 5 * Add gem matrix * Replace render :text by render :plain * Permit parameter * Update folder structure, module name to math with file name accroding to Zeitwerk rule
