+++ app/controllers/rb_releases_controller.rb
@@ -38,7 +38,7 @@ class RbReleasesController < RbApplicationController
   end
 
   def edit
-    if request.post? and @release.update_attributes(release_params)
+    if request.post? and @release.update(release_params)
       flash[:notice] = l(:notice_successful_update)
       redirect_to :controller => 'rb_releases', :action => 'show', :release_id => @release
 #    else
@@ -51,11 +51,11 @@ class RbReleasesController < RbApplicationController
     attribs = params.select{|k,v| (!except.include? k) and (RbRelease.column_names.include? k) }
     attribs = attribs.to_enum.to_h
     begin
-      result  = @release.update_attributes attribs
+      result  = @release.update attribs
     rescue => e
       Rails.logger.debug e
       Rails.logger.debug e.backtrace.join("\n")
-      render :text => e.message.blank? ? e.to_s : e.message, :status => 400
+      render :plain => e.message.blank? ? e.to_s : e.message, :status => 400
       return
     end
