+++ app/controllers/rb_releases_multiview_controller.rb
@@ -19,7 +19,7 @@ class RbReleasesMultiviewController < RbApplicationController
     if request.post?
       # Convert id's into numbers and remove blank
       params[:release_multiview][:release_ids]=selected_ids(params[:release_multiview][:release_ids])
-      @release_multiview.attributes = params[:release_multiview]
+      @release_multiview.attributes = allowed_params
 
       if @release_multiview.save
         flash[:notice] = l(:notice_successful_create)
@@ -34,7 +34,7 @@ class RbReleasesMultiviewController < RbApplicationController
       # Convert id's into numbers and remove blank
       params[:release_multiview][:release_ids]=selected_ids(params[:release_multiview][:release_ids])
 
-      if @release_multiview.update_attributes(params[:release_multiview])
+      if @release_multiview.update(allowed_params)
         flash[:notice] = l(:notice_successful_update)
         redirect_to :controller => 'rb_releases_multiview', :action => 'show', :release_multiview_id => @release_multiview
       end
@@ -47,6 +47,12 @@ class RbReleasesMultiviewController < RbApplicationController
   def destroy
     @release_multiview.destroy
     redirect_to :controller => 'rb_releases', :action => 'index', :project_id => @project
+  end
+
+  private
+
+  def allowed_params
+    params.require(:release_multiview).permit(:name, :description, {:release_ids => []})
   end
 
 end
