+++ app/controllers/rb_tasks_controller.rb
@@ -10,7 +10,7 @@ class RbTasksController < RbApplicationController
     begin
       @task  = RbTask.create_with_relationships(params, User.current.id, @project.id)
     rescue => e
-      render :text => e.message.blank? ? e.to_s : e.message, :status => 400
+      render :plain => e.message.blank? ? e.to_s : e.message, :status => 400
       return
     end
