+++ app/views/rb_releases/_sprint.html.erb
@@ -46,4 +46,4 @@
   </table>
 <% end %>
 </div>
-<%= context_menu issues_context_menu_path %>
+<%= context_menu | javascript_tag("contextMenuInit()") %>
