+++ db/migrate/039_add_show_in_scrum_stats_project_setting.rb
@@ -1,6 +1,6 @@
 class AddShowInScrumStatsProjectSetting < ActiveRecord::Migration[5.2]
   def self.up
-    add_column :rb_project_settings, :show_in_scrum_stats, :boolean, {:default => true, :null => false}
+    add_column :rb_project_settings, :show_in_scrum_stats, :boolean, default: true, null: false
   end
 
   def self.down
