+++ lib/backlogs_custom_field_patch.rb
@@ -1,19 +1,17 @@
 require_dependency 'custom_field'
 
-module Backlogs
-  module CustomFieldPatch
-    def customized_class
-      if self.respond_to?(:rb_sti_class)
-        self.rb_sti_class.customized_class
-      else
-        super
-      end
+module BacklogsCustomFieldPatch
+  def customized_class
+    if self.respond_to?(:rb_sti_class)
+      self.rb_sti_class.customized_class
+    else
+      super
     end
   end
 end
 
 class CustomField
   class << self
-    prepend Backlogs::CustomFieldPatch
+    prepend BacklogsCustomFieldPatch
   end
 end
