+++ lib/backlogs_tracker_patch.rb
@@ -1,7 +1,6 @@
 require_dependency 'user'
 
-module Backlogs
-  module TrackerPatch
+module BacklogsTrackerPatch
     def self.included(base) # :nodoc:
       base.extend(ClassMethods)
       base.send(:include, InstanceMethods)
@@ -19,7 +18,6 @@ module Backlogs
         return (issue_statuses.select{|s| !s.default_done_ratio.nil? && s.default_done_ratio < r && !s.is_closed?}.sort{|a, b| b.default_done_ratio <=> a.default_done_ratio} + [nil])[0]
       end
     end
-  end
 end
 
-Tracker.send(:include, Backlogs::TrackerPatch) unless Tracker.included_modules.include? Backlogs::TrackerPatch
+Tracker.send(:include, BacklogsTrackerPatch) unless Tracker.included_modules.include? BacklogsTrackerPatch
